+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/BinaryMessage/Bridge/KernelCommunication.BinaryMessageBridge.asmdef
@@ -1,12 +1,12 @@
 {
-    "name": "KernelCommunication.BinaryMessageProcessor",
+    "name": "KernelCommunication.BinaryMessageBridge",
     "rootNamespace": "",
     "references": [
-        "GUID:211458814b4045e7ad894e13b8f8d919",
-        "GUID:374252a1740a422eb7762fc12ff3bf8a",
-        "GUID:80f579cff5cf4136aa8969bf6fa1098c",
+        "GUID:0e5c4d710d434913bfc909adce971728",
+        "GUID:fbcc413e192ef9048811d47ab0aca0c0",
+        "GUID:97d8897529779cb49bebd400c7f402a6",
         "GUID:3047fd2f05bcaae498a99a3f9321f9a8",
-        "GUID:132013f160565434195ba5960832b346"
+        "GUID:3b80b0b562b1cbc489513f09fc1b8f69"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
